requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.49.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
astroid==3.3.9
attrs==25.3.0
audioread==3.0.1
black==25.1.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.2
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
datasets==3.5.0
decorator==5.2.1
dill==0.3.8
distro==1.9.0
docker-pycreds==0.4.0
docstring-parser==0.16
docstring-to-markdown==0.16
dotenv==0.9.9
elevenlabs==1.56.0
eval-type-backport==0.2.2
execnet==2.1.1
filelock==3.18.0
fonttools==4.57.0
frozenlist==1.6.0
fsspec==2024.12.0
gitdb==4.0.12
gitpython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.167.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.87.0
google-cloud-bigquery==3.31.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-generativeai==0.8.4
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grayswan-api==0.1.0a49
grpc-google-iam-v1==0.14.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.8
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.30.2
hydra-core>=1.3
hydra-colorlog>=1.2
idna==3.10
importlib-metadata==8.6.1
iniconfig==2.1.0
isort==6.0.1
jedi==0.19.2
jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonlines==4.0.0
kiwisolver==1.4.8
lazy-loader==0.4
librosa==0.10.2.post1
llvmlite==0.44.0
markdown-it-py==3.0.0
markupsafe==3.0.2
matplotlib==3.10.1
mccabe==0.7.0
mdurl==0.1.2
msgpack==1.1.0
multidict==6.4.3
multiprocess==0.70.16
mypy-extensions==1.0.0
narwhals==1.35.0
nodeenv==1.9.1
numba==0.61.2
numpy==2.2.4
omegaconf==2.3.0
openai==1.70.0
opencv-python==4.11.0.86
packaging==25.0
pandas==2.2.3
parso==0.8.4
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.7
plotly==6.0.1
pluggy==1.5.0
pooch==1.8.2
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pydantic==2.11.1
pydantic-core==2.33.0
pydub==0.25.1
pygments==2.19.1
pylint==3.3.6
pyparsing==3.2.3
pyright==1.1.399
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-xdist==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-io==0.3
python-lsp-jsonrpc==1.1.2
python-lsp-server==1.12.2
pytz==2025.2
pyyaml==6.0.2
redis==5.2.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rsa==4.9.1
ruff==0.11.6
safetensors==0.5.3
-e ./safety-tooling
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
sentry-sdk==2.26.1
setproctitle==1.3.5
setuptools==78.1.0
shapely==2.1.0
shellingham==1.5.4
simple-parsing==0.1.7
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
tabulate==0.9.0
tenacity==9.1.2
termcolor==3.0.1
threadpoolctl==3.6.0
tiktoken==0.9.0
together==1.5.4
tokenizers==0.21.1
tomlkit==0.13.2
tqdm==4.67.1
transformers==4.50.3
typer==0.15.2
typing-extensions==4.13.2
typing-inspection==0.4.0
tzdata==2025.2
ujson==5.10.0
uritemplate==4.1.1
urllib3==2.4.0
wandb==0.19.9
websockets==15.0.1
xxhash==3.5.0
yarl==1.20.0
zipp==3.21.0